pyproject.toml
[tool.poetry]
name = "simplejwt-demo"
version = "0.1.0"
description = ""
authors = ["James Wettenhall <James.Wettenhall@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "simplejwt_demo"}]

[tool.poetry.dependencies]
python = "^3.9"
Django = "==3.2.16"
djangorestframework = "3.14.0"
djangorestframework-simplejwt = "5.2.1"

[tool.poetry.group.test.dependencies]
pytest = "^6.0.0"
pytest-django = "*"
pytest-cov = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
